// Day16-24 Angular/Day20 CRUD using signals/user-crud/src/app/core/models/user.model.ts
export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address?: {
        street: string;
        suite: string;
        city: string;
        zipcode: string;
        geo: {
            lat: string;
            lng: string;
        };
    };
    phone?: string;
    website?: string;
    company?: {
        name: string;
        catchPhrase: string;
        bs: string;
    };
    role?: 'admin' | 'user'; // For authorization simulation
}